# web/src/lib/api/openapi.yaml
components:
...
  securitySchemes:
    apiKeyAuth:
      description: API key created via POST /api/v1/keys
      in: header
      name: X-API-Key
      type: apiKey
    bearerAuth:
      bearerFormat: composer-api-key
      description: "API key passed as `Authorization: Bearer <key>`"
      scheme: bearer
      type: http
    cookieAuth:
      description: Session cookie set by POST /api/v1/auth/login
      in: cookie
      name: composer_session
      type: apiKey
info:
  contact:
    name: Composer
    url: https://github.com/erfianugrah/composer
  description: A lightweight, self-hosted Docker Compose management platform with GitOps, pipelines, and RBAC.
  license:
    identifier: MIT
    name: MIT
  title: Composer
  version: 0.15.0
openapi: 3.1.0
paths:
...
  /api/v1/hooks/{id}:
    post:
      description: Receives push events from a configured git host (GitHub, Gitea, Forgejo). Validates the HMAC signature against the per-webhook secret and enqueues the configured sync / deploy / pipeline action. Served by a raw chi handler so the raw request body is available for signature validation; not callable from typed API clients.
      operationId: receiveWebhook
      parameters:
        - description: Webhook ID created via POST /api/v1/webhooks
          in: path
          name: id
          required: true
          schema:
            format: uuid
            type: string
      responses:
        "202":
          description: Webhook accepted; job enqueued
        "400":
          description: Malformed payload
        "401":
          description: Signature validation failed
        "404":
          description: Unknown webhook ID
      security: []
      summary: Inbound git webhook receiver
      tags:
        - webhooks
...
security:
  - cookieAuth: []
  - apiKeyAuth: []
  - bearerAuth: []
servers:
  - description: Current host
    url: /
...